third_party/blink/web_tests/external/wpt/html/semantics/document-metadata/the-meta-element/color-scheme/meta-color-scheme-remove-expected.txt

This is a testharness.js-based test.
FAIL First meta applies. assert_equals: Root element's color-scheme should be 'normal' expected "normal" but got "dark"
FAIL Second meta applies after first one is removed. assert_equals: Root element's color-scheme should be 'normal' expected "normal" but got "light"
PASS Initial color-scheme with both meta elements removed.
Harness: the test ran to completion.
